requirements.txt
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.6